Watch and build with Webpack instead of chokidar
This comes with a bit of debt. If we eventually want to build the dist
asset using CI we will need a way to split the webpack config by
environment. For the moment, it all happens locally so I think having
one webpack config stuck in "development" is fine for now.