Type checking as an explicit step in jest
Jest was doing type checking as it compiled the source using ts-jest as
a transform. I am separating the type checking step into an indepedent
runner (jest-runner-tsc) and suppressing type checking in ts-jest
(setting diagnostics to false). This way if type checking can be
considered separately from testing or linting.