ui-components package for "core components" (#2)
* Astroturf setup using Babel

Trying to layout a library structure compiled by Babel. Seems to work
with the astroturf method of defining CSS-in-JS modules, but not when
importing Sass via CSS modules.

* Build Components using Webpack and Astroturf

Attempting to build a component defining styles using astroturf's `css`
tag to extract Sass to CSS Module. Currently build works, but upon npm
link the consuming app produces an error.

* Fix webpack config so css module file is created

Updated ts rule to include js files which fixes css file not found error.
Switched export in index.js to a named export which fixed import error.
Test in astroturf-example with no changes to config.

* moving .gitignore to whole repo

* Adding .npmignore to start test publishing

* updating pre-release tag for test publishing

* Publish

 - @cockroachlabs/ui-components@0.1.0-alpha.20200207

* don't need a package lock for a lib

* moving Badge to css module

* add repo to package.json

* a very simple test (setting up jest)

* adding enzyme and testing badge props

* ignore TODO files

* this webpack bar is neat

* using Sample instead of Badge

* cutting "tokens" for now, will include with production ready components

* polishing npmignore

* better test flow

I think it is more common to have this as a dev setup

* Watch and build with Webpack instead of chokidar

This comes with a bit of debt. If we eventually want to build the dist
asset using CI we will need a way to split the webpack config by
environment. For the moment, it all happens locally so I think having
one webpack config stuck in "development" is fine for now.

* adding additional authors

* example READMEs for documentation

* publish public packages only

* Testing prerelease publishing

As I'm testing out publishing, I want to use prerelease versions and I
don't think the _alpha_, _beta_, _rc_, conventions used by CRDB apply
here. If prerelease versions need to happen, let's just use "prerelease"
as the prerelease id for simplicity.

* Declaring peerDependencies

- Adding `react` and `react-dom` as peerDependencies
- Generating the package-lock.json (and changing the .npmrc to keep it)
  since we will need it for CI and stuff

* cutting lodash/isUndefined

* Adding linting

Adding linting with the ESlint and Prettier (using the same
configuration as CockroachCloud) and additionally running linting as a
part of jest using `jest-runner-eslint`.

* Type checking as an explicit step in jest

Jest was doing type checking as it compiled the source using ts-jest as
a transform. I am separating the type checking step into an indepedent
runner (jest-runner-tsc) and suppressing type checking in ts-jest
(setting diagnostics to false). This way if type checking can be
considered separately from testing or linting.

* Nevermind

I take it all back. jest-runner-tsc seems to take issue with CSS
Modules. Perhaps a task for another day.

* defaultProps are getting deprecated

Co-authored-by: Lauren Barker <ms.lebarker@gmail.com>