Storybook setup for UI Components
This is a pretty minimal setup on which stories can be build on top of.
It creates a simple placeholder theme, installs a few addons, and has a
simple utility for adding sugar to the knobs addon. All to indicate a
direction that may be used as a pattern for future storybook
installations.