Badge Component Story (#7)
* adding snapshots and tokens.scss

* Storybook setup for UI Components

This is a pretty minimal setup on which stories can be build on top of.
It creates a simple placeholder theme, installs a few addons, and has a
simple utility for adding sugar to the knobs addon. All to indicate a
direction that may be used as a pattern for future storybook
installations.

* Adding a story for the Badge Component

This is just a start to fully document the Badge Component. This story
was created for the purpose of demonstrating how the component
changes when props are altered.

* providing example story, refactoring badge

* killing welcome story

* messing with static storybook generation

* I don't think we want to publish storybook package