Adjusting PR github action
I'm renaming the action to be a little more descriptive and adding more
steps with names to make this action a little easier to see what's going
on.

I have added a step to build each pacakge after bootstrap. This will
prevent errors where a package is linked, but the dist/ folder may not
exist and so the package will appear to not exist or not be installed
correctly.