Building Typescript declarations with tsc
Currently the ui-components package is authored in Typescript and then
transpiled to JavaScript using Babel with the TypeScript preset. This
has two effects: first, that the code is never type checked (we leave
that to the IDE environment) and second, declarations are not generated
for the transpiled code. This results in functioning package code that
can't be typed by the consuming application.

In an effort to correct this I have taken a few steps.

- the extensions of index files have been changed from `js` to `ts`.
This will aid the TypeScript compiler in processing them without having to
explicitly be told to do so.

- adding a tsconfig.json that will generate declarations
to this dist directory and then identifying them in the package.json
(with the `types` option).

- I've added an script entry to run the TypeScript compiler as well
as a lifecylce script to generate the declarations on prepublish and install.