Correcting type errors
Now that type checking is working correctly, I am making some
adjustments to the component and story code to remove type errors and
warnings.

One note on a change in convention, is that I am now exporting all types
along with the component from the main component file. To simplify the
levels of `index.ts` files I have removed the index from `Badge` and
`Avatar` for simplicity. Notice the index is still there for `Input` as
it is a component complex enough to leverage it.