Just kidding, we need prepublishOnly
We need to have prepublishOnly on ui-components. Here's what I think is
happening; when we publish a change to ui-components or on any of its
dependencies lerna will look at the dependency tree and execute the
prepublish step on ui-components AND on its dependencies?

Publishing ui-components without a prepublish step, resulting in no
build happening.