Correcting fonts in stories
Fonts were loading correctly, but were not being applied to stories via
the "cockroach theme" (because that's not how the theme works). I think
there had been an attempt to apply fonts to the stories by using a
layout components, but had incorrectly used OpenSans instead of the
correct Source Sans Pro.

I am correcting the font and applying the layout component to all of the
current stories. There's also a lot of "prettifying" that happened in the
editor.