Add onSortTable handler for StatementsPage
When StatementsPage component was extracted, analytics tracking
functionality which was embedded into component has been removed and
should be handled outside of component (with respective callback
functions).
`trackTableSort` tracking function was missed. To make it work,
`onSortingChange` prop is added to handle it outside.

Also unnecessary `refreshDiagnosticsReports` is removed because it
duplicates already existing `refreshStatementDiagnosticsRequests` prop.