Extend ColumnDescriptor with column name field
ColumnDescriptor configuration provides information about existing
columns in table. The issue appeared when user sorts data by some
column and it is necessary to know column name (to send it as part
of analytics data). Previously, `title` column was used for this
purpose, but the problem is that `title` field is ReactNode, not string.
It means, after casting field to string - it always has [object Object]
value instead of meaningful column name.

With current change, ColumnDescriptor configuration is extended with
extra string field "name" as unique column name.