Debugging workflows
The original goal of this work was to prevent workflows from running on
the master branch of forks which are expected to fail and create noise.

I am also trying to fix an npm publishing workflow that has never worked
because of the was that the master branch is protected. I may have to
remove "protected" status from the master branch to get this to work.
Another issue with the publishing workflow is that it attempts to
publish a prerelease version, but tags this version as "latest" which is
not what we want. I am adding a preid to the prerelease version (so the
version will be vX.X.X-alpha.X instead of vX.X.X-X) which can more
easily distinguish it from a stable version.

Doing this I thought about our dist-tags for ui-components and have
decided to try "latest" as versions published by lerna, "next" as
versions published after a PR is merged, and "testing" as development
versions published manually for testing in applications (where
symlinking may be cumbersome). To make this a little easier I have added
an npm script called "publish-test-version" to run these commands.