Call backend API from Storybook
The main idea is to enable testing components with backend integration,
using real data. It allows testing of components connected to redux
store and perform some kind of integration testing.

To establish communication, proxy-serve is provided as part of
Storybook setup and it proxies API request to target url which
has to be provided as ENV variable:
```
TARGET=localhost:56789 yarn storybook
```

`demoFetch` component and storybook are for demo purposes only
and can be removed lately.