Icon updates (#120)
* Adding resolution for prismjs (#117)

Adding resolution for prismjs@1.21.0 to resolve https://github.com/advisories/GHSA-wvhm-4hhf-97x9

* Icon files update to latest Figma version

Updating icon svg files that have been uniformly sized. There were some
icons that had name changes and so file and export names have been
updated.

* update Icon story description

* Updating Icon sizes to match Figma

Sizes were taken from original Icon component in Cloud Console. Updating
to match design in Figma.

* Storybook using ui-components@next

I'm changing the version of ui-components in storybook to use the npm
dist-tag next so that the deployed version of storybook will match the
version of components that are published as they are merged.

This might be reckless, but it dawns on me that we deploy Storybook when
we merge a PR to master and so it should use the version of
ui-components that is also deployed when merged to master.

This might get messed up the next time we `learna publish` anyway.

* Configuring SVGR to remove fill attributes

SVGr uses the SVGO optimizer under the hood. I'm adding a configuration
to remove the "fill" attribute in the svgs exported from Figma, allowing
the fill to be applied by css.

* Adding prepare script to build icons

Adding a lifecycle script to build icons when running `npm run init` at
the root of the project (which runs `lerna bootstrap`).

* Updating cirle icons with white paths

Some icons had paths with white fills laid over black fills. When all
fill attributes are removed, this was resulting in solid circles. SVGs
were updated in Figma and re-imported to be transparent intersections
instead.