ui-componets: Extending Button with 'as' prop
We have decided that we will continue the pattern of using Buttons as
navigational components. To make this obfuscation semantically correct,
we need to render a navigational element when using Button to ...
navigate. In practice this means when `<Button>` is used as an external
link, then it will render as an `<a>`. When `<Button>` is used as an
internal link, the application should have the ability to specify a
routing component like `<Link>` from `react-router`.

To solve this issue, I am extending Button with an `as` prop, to allow
the user to define what component the Buttons should render _as_. This
is encouraged behavior in React, but this creates some difficulty in
determining the appropriate props for the specific type of element that
Button will render as. I wanted to maintain the pattern of combining
specified props unique to the component with native props of the rendered
element, and so I have employed TypeScript generics to do this in a
dynamic way with a specified element from the `as` prop.