ui-component: Button style updates and as support
Adjusting some of the styles for Button to conform to a design review
(see issue #123). These include a change to font-size, font-weight,
font-family, and letter-spacing. Referencing the Figma file, I'm also
making an adjustment to line-height.

In doing some trial integrations into Cockroach Cloud Console, we assume
that Buttons will behave with their native display of `inline-block`
and so I am setting the display to be `inline-flex` rather than `flex`.

To support the `as` prop (with the assumption that uses will use Buttons
as `<a>`) I am applying some styles to bring _links_ into visual
conformity with buttons. Removing default text decoration, and forcing
box sizing _border-box_ to have a consistent element height.