fix Button component styles
Buttons that rendered with icons didn't
set correct icon style depending on
button type and state.

Now, color of icon is updated in the same
way as Button colors.

In addition `icon` prop is changed to accept
`ReactNode` now, because function call with
returning ReactNode is redundant in current
component.