rewrite Dropdown component with new designs
The main reason to rewrite entirely Dropdown component is
that it was very difficult to maintain, extend to current
needs. Lots of custom logic was added inside so it felt like
very custom component.

Current component is designed to easily customize it by
overriding values via props and predefined options satisfy
custom designs.