update crdb protobuf dependency (#32)
This change introduces an updated Statements Response
object which uses Long IDs instead of strings.

The Transactions Page fixture has been updated to be
properly typed.

One major logic change is that we need to use the `eq`
method of the `Long` type in order to filter for
matching statements for a given transactions. Previously
because this was a string `===` would suffice.