sortedtable refactoring: remove sortabletable, simplify sortedtable
Originally we had 2 components for table functionality sortabletable and sortedtable. sortabletable contains mostly all code, sortedtable is just wrap it and adds couple additional api's. Through admin-ui codebase everywhere used wrapper version of the table. Only in one, seems to be very old, part of the code is used the table without wrapper.

In order to have single version of it, I moved all code in one place, under sortedtable component, as it is used everywhere. Another issue was that all code lived in one file, so tried to split it in smaller components (, , ) with corresponding css in modular style.

Overall types\props consumed by refactored table not changed, so integrate it in admin-ui should be as easy as just change import statements.

And also this pr contains latests updates from CRDB-Design-System_V2 (fonts, colors, paddings).