use yarn workspaces
To avoid the issue when more than one instance of React library
is used during development, yarn workspaces is used.
It hoists all common dependencies to the top level node_modules
and as result every package relies on the same instance of React
library and others.

In addition, react version mismatch was caused by "react-final-form"
library which in addition required the latest version (ver 17), to
solve this and similar issues - react specific versions are added to
"resolutions" section to force nested dependencies to use specific
version.