transactions page: aggregate by fingerprint
This change introduces some helper functions that let us aggregate
transaction stats returned by the DB by the fingerprint and app. When
CRDB returns transaction stats there may be copies of the transaction
across different node_ids which executed them. In the UI we don't
present that distinction in the transaction list so, like we do with
statements, we need to aggregate stats across copies of the same
fingerprint.