transactions page: aggregate by fingerprint  (#166)
This change introduces some helper functions that let us aggregate
transaction stats returned by the DB by the fingerprint and app. When
CRDB returns transaction stats there may be copies of the transaction
across different node_ids which executed them. In the UI we don't
present that distinction in the transaction list so, like we do with
statements, we need to aggregate stats across copies of the same
fingerprint.

* rename addNumericStats to aggregateNumericStats
* merge util.spec.ts into utils.spec.ts

* filter transactions by app using exact matches

Previously we were filtering using prefix matches for app names
since that's necessary when filter internal apps. This change
modifies the transaction filtering to disambiguate between the
prefix filtering using for internal app names and the exact match
method used for customer-set app names.