cluster-ui: fix link on statement diagnostics bundle
Before, cluster-ui base path was defined as '/' root path and
links relied on it. But now, base pack can be set to any other
path and all requests on server should append updated base path.
`getBasePath` function returns current base path or empty string
if it is not set.

This change updates the link to use `getBasePath` and build
valid URL which includes base path.