cluster-ui: Transactions Page story tests
This PR demonstrates how storybook stories
can be used as simple enzyme tests to ensure
that we don't break renders of any of our cases.

Additionally, the transaction statistics fixture
object was extracted for easier sharing and also
uses the `Require` generic type to ensure that
any future additions to the stats object will
be required to be added here.