storybook: link ui-components dependencies
`storybook-ui-components` package always depends on the latest
version of `ui-components` package and during development Storybook
has to apply all local changes.
To achieve this, `storybook-ui-components` installs `ui-components`
package as a linked dependency so it always has latest version.

Before, it relied on "next" tag that didn't work as expected and
it was a reason when local installs failed.