ui-components: remove enzyme and refactor tests to use RTL
We've decided to move away from Enzyme and HTML snapshot tests in
favor of React Testing Library and testing what the user sees. Enzyme
has been removed from the `managed-service` repo, but we are still
working to switch over in the `cockroach` repo. Removing enzyme from
the `ui` repo gets us a step closer to a unified testing front.

Since these components are generic, reusable components, the tests will
be short if we aren't testing implementation details. Some of the
refactored tests still look for class names and attributes, but we could
choose to remove these tests entirely if we aren't getting value out of
them.

Any snapshot tests were removed and not replaced.