docs: add directory for frontend-focused RFCs
The frontend-focused RFC directory is a space for frontend contributors
to propose and discuss best practices that will impact multiple teams.
We've been having discussions in #frontend-learners, but slack threads
aren't ideal for coming to a consensus, especially across teams.

The frontend-focused RFC directory is not intended for consumer-facing
features. Those RFCs should be added to `docs/RFCS` in the relevant
repository.

Release note: None