eslint-config, ui-components: bump eslint versions and fix errors
`prettier/@typescript-eslint` is now included in
`plugin:prettier/recommended`, so that extension was removed.

`@typescript-eslint/no-unused-vars` has been updated to be an error to
match the manage-service config.

The `--format=codeframe` option was removed from the lint script in
`ui-components` because that option is no longer included in ESLint.

`jest-runner-eslint.config.js` was removed because it was setting a
format option that is no longer included in ESLint core.