Test the application of expected classes (#430)
* WIP: Attempting to test Color in React Testing Library

I thought that I could simply compare the color of the rendered
component to the expected color of our design tokens. The issue
that I'm currently stuck on is that Jest won't transpile any
dependencies (meaning that if I import the existing tokens Jest will
fail, not recognizing `export`). If I compile the tokens to a CommonJS
module then TypeScript will fail the test saying there are no type
definitions.

Beyond that, is this even a strategy that we like? How else could be
replace a test for appropriate intent?

* ui components: Reverting to testing applied classes for intents

Testing that classes are applied for intent. Visual testing will
require something to take visual snapshots and do comparisons. To be
explored in another PR.