ui-components: add recharts and bar chart component to library
Currently `uPlot` is the graphing library of choice for timeseries data
across Cockroach Labs. This libraries is heavilily optimized to high
volume timeseries data, which it does great at! `uPlot` can effectively
render tens of thousands of points in a very quick timeframe.

Unfortunately! `uPlot`'s support for plugins leaves a fair bit to be
desired in terms of flexibility in visualizing categorical data. To aid
this, this PR proposes using `Recharts` with a wrapper for defining a
similarly styled graph. The first component introduced is a normal
`BarChart` component, which provides a few options for fitting an
internal use case that we have. Additional features can easily be added
through updating this component in the `ui` library.

This components functionality and a small use case is documented in the
`Charts/README.md` file, so take a look there to see how it works! This
will be an iterative process as we start to nail down what we'd like to
use.

Quick example:
```
 <BarChart
   width={400}
   height={400}
   margin={{ top: 32, left: 32, right: 32, bottom: 16 }}
   alignment="vertical"
   data={sampleData}
   mainKey="name"
   dataKeys={["count"]}
   barColors={["#ff0000", "#0000ff"]}
/>
```

Release note: Added Rechart support for the `ui-components` package.