Removing @types/react version ^18.0.0
Some monster in our transitive dependencies specified a dependency on
`@types/react@*` which hoisted version 18.0.15 into the workspace
dependencies. This (somehow) caused a type error when building
ui-components. I have added a resolution to the workspace to only
include version 17 of react types, as everything in this repo should use
React 17.