ui-components: Fixing type compatibility in InlineAlerts
The intent prop (of type InlineAlertIntent) was being passed directly to
the Icon component's fill prop (of type IconFill). InlineAlertIntent is
compatible as a subset of IconFill (which is a union of IconIntent and
some string types), but are clearly different types and not
interchangable.

The fix is to cast the InlineAlertIntent to a string and then pass the
string cast as an IconFill.