eslint-config: warn console.log statements when committed
We're seeing a decent amount of PR reviews comment on stray console.logs. We can
just stop that at the source with an eslint configuration.

Setting this to `warn` to not prevent or halt any builds upstream if DBConsole
has any console.logs in the repo. IIRC we cleared it for cloud a bit ago.

Release note: None