ui-components: upgrade to css-loader@6.8.1
packages/ui-components/ previously used css-loader explicitly but didn't
declare it as a dependency in its package.json. That dependency was
resolved to version 3.6.0 via the dependency chain:

./packages/storybook-ui-components/
  @storybook/react
    @storybook/core
      @storybook/core-server
        @storybook/builder-webpack4
          css-loader

Upgrading to storybook 7 would change how css-loader is resolved and
cause an unintentional breaking major-version upgrade. Declare the
css-loader dependency in packages/ui-components explicitly, and upgrade
to version 6.