ui-components, icons, release.yml: add Italy, bump package version, fix release scripts (#548)
* ui-components, icons, release.yml: add Italy, bump package version, fix release scripts

We are adding support for Italy as a region, so we need to add the flag. Bumping the version to release all new flag additions.

Bump versions of actions in `release.yml` so that they all support node v20. See https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/ for more details. This allows the release script to run without failures.

Also bumped the version of `auto` to fix:
```
> auto shipit

fatal: ' 9007199254740991': not an integer

Error: fatal: ' 9007199254740991': not an integer
```